+++ server/package.json
@@ -25,7 +25,8 @@
     "mongoose": "^5.5.6",
     "mongoose-unique-validator": "^2.0.3",
     "multer": "^1.4.1",
-    "nodemailer": "^6.2.1"
+    "nodemailer": "^6.2.1",
+    "redis": "^2.8.0"
   },
   "devDependencies": {
     "chai-http": "^4.3.0",
